+++ packages/dash-core-components/webpack.config.js
@@ -66,6 +66,9 @@ module.exports = (env, argv) => {
                     use: [
                         {
                             loader: 'style-loader',
+                            options: {
+                                insertAt: 'top'
+                            }
                         },
                         {
                             loader: 'css-loader',