+++ packages/dash-core-components/CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.38.1] - 2018-11-21
 ### Fixed
 - Refactored the way the Graph component would generate an unique id if none provided.
+- Default CSS imported via `style-loader` is now placed at top, so that user supplied CSS can overwrite it, fixes [#380](https://github.com/plotly/dash-core-components/issues/380)
 
 ## [0.38.0] - 2018-11-07
 ### Fixed